% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_xaringan_media_calls.R
\name{ds_xaringan_media_calls}
\alias{ds_xaringan_media_calls}
\title{Create list of all media calls in a xaringan slide deck}
\usage{
ds_xaringan_media_calls(rmd_file)
}
\arguments{
\item{rmd_file}{Path to the Rmarkdown file to be processed}
}
\value{
List of all media calls in markdown file and printed to console
}
\description{
Create list of all media calls in a xaringan slide deck
}
\examples{
ds_xaringan_media_calls("unique-is-more-important.Rmd")

}
